style: 合并多余实体

+++ themis-domain/src/main/java/com/achobeta/themis/domain/user/model/vo/UserFavoriteRegulationVO.java
@@ -41,7 +41,7 @@ public class UserFavoriteRegulationVO {
     /**
      * 法律分类ID
      */
-    private Integer lawCategoryId;
+    private Long lawCategoryId;
 
     /**
      * 发布年月日
