+++ themis-infrastructure/src/main/java/com/achobeta/themis/infrastructure/user/repo/LawDataImportRepository.java
@@ -43,7 +43,7 @@ public class LawDataImportRepository implements ILawDataImportRepository {
         }
         
         // 2. 创建分类ID到名称的映射
-        Map<Integer, LawCategory> categoryMap = new HashMap<>();
+        Map<Long, LawCategory> categoryMap = new HashMap<>();
         for (LawCategory category : categories) {
             categoryMap.put(category.getLawId(), category);
         }
